requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==4.4.2
filelock==3.16.1
fsspec==2024.10.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
openai-whisper==20240930
pillow==11.0.0
proglog==0.1.10
regex==2024.9.11
requests==2.32.3
sympy==1.13.3
tiktoken==0.8.0
torch==2.2.2
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
